+++ CHANGELOG.md
@@ -31,6 +31,8 @@
 - Fix [#88](https://github.com/znamlab/flexiznam/issues/88). Now make attributes JSON
   compatible before uploading to flexilims. This will replace special characters in
   attribute names by `_` in the database.
+- Fix [[#102](https://github.com/znamlab/flexiznam/issues/102). `add_mouse` now works
+  with mice that have special character in their allele.
 - `add_recording` and `add_sample` add the value online with the full name (including
   genealogy) rather than the short name.